fix: 헤더 컴포넌트 분리
